requirements.txt
# Install PyTorch with CUDA support first:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

faster-whisper>=0.9.0
pyannote-audio>=3.1.0
# torch>=2.0.0  # Install separately with CUDA support
# torchaudio>=2.0.0  # Install separately with CUDA support
numpy>=1.24.0
scipy>=1.10.0
librosa>=0.10.0
soundfile>=0.12.0
webrtcvad>=2.0.10
Flask
werkzeug
pyannote.audio
python-dotenv
